requirements.txt
notebook==6.4.12
torch==1.11.0
transformers==4.20.0
pandas==1.4.2
datasets==2.3.2
seaborn==0.11.2
scikit-learn==1.1.1
python-dotenv==0.20.0
wandb==0.12.21
evaluate==0.2.0
openai==0.22.0

